Implement the Networking Layer
This implements all of the models and networking required for POSTing a Metric Event to AppFit. This allows us to dynamically inject the raw event along with the project id and api key for the API Client.

This includes documentation as well as unit tests to help verify things are flowing correctly.