AppFitConfiguration Implementation
This timplements the base properties on the AppFitConfiguration. This allows us to give it a base64 encoded API Key that we currently use for all other API Services, and we can decode it to get the projectId and the apiKeyId. This is neded to build the basicAuth header as well as reference the ProjectId for each metric event.