Clean up the entire event layer
This cleans up and refactors some of the logic around taking an `AppFitEvent` and creating a raw metric event for the backend. This allows us to create a private method, vs exposing the entension to the public. This allows us to keep some of these classes more public and not userfacing, private.