Event Digester Foundation
This implements the basis of the event digester to support caching and posting the events to the network. This includes loose logic for verifying an event exists and removing it, then readding it in case something was changed. Since all events are id-tagged, this eliminates the ability to overwrite an event that might be pending.